mcp-servers/gemini-test-agent/requirements.txt
mcp>=1.0.0
google-generativeai>=0.8.0
asyncio
typing
logging